requirements.txt
flask==1.0
requests==2.19.1
pyyaml==5.1
jinja2==2.10
urllib3==1.24.1
